+++ kiselev.sergey/T2/data.cpp
@@ -3,31 +3,7 @@
 #include <ios>
 #include <iostream>
 #include "scopeGuard.hpp"
-
-std::istream& kiselev::operator>>(std::istream& input, DelimeterIO&& dest)
-{
-  std::istream::sentry sentry(input);
-  if (!sentry)
-  {
-    return input;
-  }
-  char c;
-  input >> c;
-  if (input && (c != dest.exp))
-  {
-    input.setstate(std::ios::failbit);
-  }
-  return input;
-}
-
-std::istream& kiselev::operator>>(std::istream& input, DelimetersIO&& dest)
-{
-  for (char c: dest.exp)
-  {
-    input >> DelimeterIO{ c };
-  }
-  return input;
-}
+#include "delimeters.hpp"
 
 std::istream& kiselev::operator>>(std::istream& input, UllIO&& dest)
 {
