+++ kiselev.sergey/common/delimeters.cpp
@@ -1,0 +1,27 @@
+#include "delimeters.hpp"
+#include <istream>
+
+std::istream& kiselev::operator>>(std::istream& input, DelimeterIO&& dest)
+{
+  std::istream::sentry sentry(input);
+  if (!sentry)
+  {
+    return input;
+  }
+  char c;
+  input >> c;
+  if (input && (c != dest.exp))
+  {
+    input.setstate(std::ios::failbit);
+  }
+  return input;
+}
+
+std::istream& kiselev::operator>>(std::istream& input, DelimetersIO&& dest)
+{
+  for (char c: dest.exp)
+  {
+    input >> DelimeterIO{ c };
+  }
+  return input;
+}
