+++ kiselev.sergey/common/scopeGuard.cpp
