+++ kiselev.sergey/common/scopeGuard.hpp
